moved some methods to base
