+++ src/nemos/glm/glm.py
@@ -4,7 +4,6 @@
 from __future__ import annotations
 
 import warnings
-from pathlib import Path
 from typing import Any, Callable, Literal, Optional, Union
 
 import equinox as eqx
@@ -271,11 +270,6 @@ class GLM(BaseGLM[GLMUserParams, GLMParams]):
                     f"``{correct_model}`` object."
                 )
             raise TypeError(error_msg)
-
-    def __sklearn_clone__(self) -> "GLM":
-        """Clone the GLM."""
-        params = self.get_params(deep=False)
-        return self.__class__(**params)
 
     def _predict(
         self, params: GLMParams, X: Union[dict[str, jnp.ndarray], jnp.ndarray]
@@ -742,81 +736,6 @@ class GLM(BaseGLM[GLMUserParams, GLMParams]):
         )
 
         return self._validator.from_model_params(updated_params), updated_state
-
-    def save_params(self, filename: Union[str, Path]):
-        """
-        Save GLM model parameters to a .npz file.
-
-        This method allows to reuse the model parameters. The saved parameters can be loaded back
-        into a GLM instance using the `load_params` function.
-
-        Parameters
-        ----------
-        filename :
-            The name of the file where the model parameters will be saved. The file will be saved in `.npz` format.
-
-        Examples
-        --------
-        >>> import nemos as nmo
-        >>> # Create a GLM model with specified parameters
-        >>> solver_args = {"stepsize": 0.1, "maxiter": 1000, "tol": 1e-6}
-        >>> model = nmo.glm.GLM(
-        ...     regularizer="Ridge",
-        ...     regularizer_strength=0.1,
-        ...     observation_model="Gamma",
-        ...     solver_name="BFGS",
-        ...     solver_kwargs=solver_args,
-        ... )
-        >>> for key, value in model.get_params().items():
-        ...     print(f"{key}: {value}")
-        inverse_link_function: <function one_over_x at ...>
-        observation_model: GammaObservations()
-        regularizer: Ridge()
-        regularizer_strength: 0.1...
-        solver_kwargs: {'stepsize': 0.1, 'maxiter': 1000, 'tol': 1e-06}
-        solver_name: BFGS[...]
-        >>> # Save the model parameters to a file
-        >>> model.save_params("model_params.npz")
-        >>> # Load the model from the saved file
-        >>> model = nmo.load_model("model_params.npz")
-        >>> # Model has the same parameters before and after load
-        >>> for key, value in model.get_params().items():
-        ...     print(f"{key}: {value}")
-        inverse_link_function: <function one_over_x at ...>
-        observation_model: GammaObservations()
-        regularizer: Ridge()
-        regularizer_strength: 0.1...
-        solver_kwargs: {'stepsize': 0.1, 'maxiter': 1000, 'tol': 1e-06}
-        solver_name: BFGS[...]
-
-        >>> # Saving and loading a custom inverse link function
-        >>> model = nmo.glm.GLM(
-        ...     observation_model="Poisson",
-        ...     inverse_link_function=lambda x: x**2
-        ... )
-        >>> model.save_params("model_params.npz")
-        >>> # Provide a mapping for the custom link function when loading.
-        >>> mapping_dict = {
-        ...     "inverse_link_function": lambda x: x**2,
-        ... }
-        >>> loaded_model = nmo.load_model("model_params.npz", mapping_dict=mapping_dict)
-        >>> # Now the loaded model will have the updated solver_name and solver_kwargs
-        >>> for key, value in loaded_model.get_params().items():
-        ...     print(f"{key}: {value}")
-        inverse_link_function: <function <lambda> at ...>
-        observation_model: PoissonObservations()
-        regularizer: UnRegularized()
-        regularizer_strength: None
-        solver_kwargs: {}
-        solver_name: GradientDescent[...]
-        """
-
-        # initialize saving dictionary
-        fit_attrs = self._get_fit_state()
-        fit_attrs.pop("solver_state_")
-        string_attrs = ["inverse_link_function"]
-
-        self._save_params(filename, fit_attrs, string_attrs)
 
 
 class PopulationGLM(GLM):
